+++ chromium/recorder_bridge/recorder_switches.h
@@ -13,6 +13,8 @@ inline constexpr char kChildBootstrapMetadataEnvironment[] =
     "A11Y_RECORDER_CHILD_BOOTSTRAP_METADATA";
 inline constexpr wchar_t kChildBootstrapMetadataEnvironmentWide[] =
     L"A11Y_RECORDER_CHILD_BOOTSTRAP_METADATA";
+inline constexpr wchar_t kBridgeLogFileEnvironmentWide[] =
+    L"A11Y_RECORDER_BRIDGE_LOG_FILE";
 
 }  // namespace a11y_recorder